move ProcessingResult
